requirements.txt
absl-py==1.0.0
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.27
decorator==5.1.1
en==0.0.1
et-xmlfile==1.1.0
executing==0.8.2
filelock==3.4.2
flatbuffers==2.0
fonttools==4.29.1
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.3
ipython==7.6.0
jedi==0.18.1
joblib==1.1.0
kiwisolver==1.3.2
kmeans-pytorch==0.3
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.22.2
oauthlib==3.2.0
onnx==1.10.2
onnxruntime==1.10.0
opencv-python==4.5.5.62
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
prettytable==3.0.0
prompt-toolkit==2.0.10
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools @ git+https://github.com/cocodataset/cocoapi.git@8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9#subdirectory=PythonAPI
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
stack-data==0.1.4
submitit==1.4.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
timm==0.5.4
tokenizers==0.10.3
tomli==2.0.0
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.62.3
traitlets==5.1.1
transformers==4.5.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
xmltodict==0.12.0
zipp==3.7.0
